+++ pocsuite3/lib/request/patch/__init__.py
@@ -4,6 +4,7 @@ from .hook_request import patch_session
 from .add_httpraw import patch_addraw
 from .hook_request_redirect import patch_redirect
 
+
 def patch_all():
     disable_warnings()
     remove_ssl_verify()