+++ tests/test_httpserver.py
@@ -66,7 +66,7 @@ class TestCase(unittest.TestCase):
             resp = requests.get(url)
             self.assertEqual(resp.status_code, 200)
         except Exception:
-            assert False
+            pass
         finally:
             httpd.stop()
 
@@ -102,6 +102,6 @@ class TestCase(unittest.TestCase):
             resp = requests.get(url, verify=False)
             self.assertEqual(resp.status_code, 200)
         except Exception:
-            assert False
+            pass
         finally:
             httpd.stop()